requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
pushbullet.py==0.12.0
twilio==8.10.0
schedule==1.2.0
python-dotenv==1.0.0
pydantic==2.5.0
